Add option for parameters
Parameters will be needed for the second part of the API token request process but they must be optional since the first part does not require any parameters. They will also be needed when making certain API calls such as onces that modify.

The splat lines were added so that the vanilla oauth method would be sent if no parameters were used. Plus splatting is just cleaner in general.