Initial version
The goal of this function will be to automate most if not all of the OAuth token request process.

By default, the token, token secret, consumer key, and consumer key secret will be stored as secure strings in text files to make future calls easier. The DoNotStore switch can be used to not store the keys and tokens and instead only output them to the screen.